# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dispatch — iMessage Dashboard API
 * OpenAPI spec version: 0.2.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  timestamp: string;
}

export type DashboardStatsAgentStatus = typeof DashboardStatsAgentStatus[keyof typeof DashboardStatsAgentStatus];


export const DashboardStatsAgentStatus = {
  online: 'online',
  offline: 'offline',
  unknown: 'unknown',
} as const;

export type MessageStatus = typeof MessageStatus[keyof typeof MessageStatus];


export const MessageStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface Message {
  id: number;
  phoneNumber: string;
  content: string;
  status: MessageStatus;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  duration?: number | null;
  retryCount: number;
  sentAt: string;
}

export interface DashboardStats {
  agentStatus: DashboardStatsAgentStatus;
  agentConnected: boolean;
  messagesReady: boolean;
  appleScriptReady: boolean;
  connectedAccounts: number;
  connectedDevices: number;
  sentToday: number;
  failedToday: number;
  pendingMessages: number;
  totalMessages: number;
  /** @nullable */
  avgSendTimeMs?: number | null;
  /** @nullable */
  lastHeartbeat?: string | null;
  activeAgents: number;
  recentMessages: Message[];
}

export interface MessagePage {
  messages: Message[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface MessageInput {
  /** @minItems 1 */
  phoneNumbers: string[];
  /** @minLength 1 */
  content: string;
}

export interface SendResult {
  sent: number;
  failed: number;
  total: number;
  results: Message[];
}

export interface SettingsMap {[key: string]: string | null}

export type MacAgentStatusProperty = typeof MacAgentStatusProperty[keyof typeof MacAgentStatusProperty];


export const MacAgentStatusProperty = {
  online: 'online',
  offline: 'offline',
} as const;

export interface MacAgent {
  id: number;
  agentId: string;
  hostname: string;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  macosVersion?: string | null;
  /** @nullable */
  nodeVersion?: string | null;
  /** @nullable */
  agentVersion?: string | null;
  /** @nullable */
  messagesAppRunning?: boolean | null;
  /** @nullable */
  messagesAppAvailable?: boolean | null;
  /** @nullable */
  appleScriptAvailable?: boolean | null;
  connectedAccounts?: string[] | null;
  connectedDevices?: string[] | null;
  /** @nullable */
  latencyMs?: number | null;
  /** @nullable */
  cpuUsage?: number | null;
  /** @nullable */
  memoryUsage?: number | null;
  queueSize: number;
  /** @nullable */
  lastHeartbeatAt?: string | null;
  /** @nullable */
  lastActivityAt?: string | null;
  status: MacAgentStatusProperty;
  /** @nullable */
  macAgentUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface HeartbeatInput {
  agentId: string;
  hostname: string;
  os: string;
  /** @nullable */
  macosVersion?: string | null;
  nodeVersion: string;
  agentVersion: string;
  messagesAppRunning: boolean;
  messagesAppAvailable: boolean;
  appleScriptAvailable: boolean;
  connectedAccounts: string[];
  connectedDevices: string[];
  /** @nullable */
  latencyMs?: number | null;
  cpuUsage: number;
  memoryUsage: number;
  queueSize: number;
  /** @nullable */
  lastActivityAt?: string | null;
}

export interface MacAgentStatus {
  connected: boolean;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  isMac?: boolean | null;
  /** @nullable */
  messagesAppAvailable?: boolean | null;
  /** @nullable */
  appleScriptAvailable?: boolean | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  latencyMs?: number | null;
  /** @nullable */
  hostname?: string | null;
  /** @nullable */
  agentVersion?: string | null;
  connectedAccounts?: string[] | null;
  connectedDevices?: string[] | null;
  /** @nullable */
  lastHeartbeat?: string | null;
}

export type DeviceService = typeof DeviceService[keyof typeof DeviceService];


export const DeviceService = {
  iMessage: 'iMessage',
  SMS: 'SMS',
} as const;

export interface Device {
  id: number;
  agentId: string;
  displayName: string;
  service: DeviceService;
  /** @nullable */
  accountId?: string | null;
  available?: boolean;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
}

export type LogEntryLevel = typeof LogEntryLevel[keyof typeof LogEntryLevel];


export const LogEntryLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface LogEntry {
  id: number;
  level: LogEntryLevel;
  category: string;
  message: string;
  /** @nullable */
  data?: string | null;
  /** @nullable */
  agentId?: string | null;
  createdAt: string;
}

export interface LogPage {
  logs: LogEntry[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type GetMessagesParams = {
status?: GetMessagesStatus;
search?: string;
page?: number;
pageSize?: number;
sort?: GetMessagesSort;
};

export type GetMessagesStatus = typeof GetMessagesStatus[keyof typeof GetMessagesStatus];


export const GetMessagesStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export type GetMessagesSort = typeof GetMessagesSort[keyof typeof GetMessagesSort];


export const GetMessagesSort = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ExportMessagesParams = {
format: ExportMessagesFormat;
status?: ExportMessagesStatus;
};

export type ExportMessagesFormat = typeof ExportMessagesFormat[keyof typeof ExportMessagesFormat];


export const ExportMessagesFormat = {
  csv: 'csv',
  json: 'json',
} as const;

export type ExportMessagesStatus = typeof ExportMessagesStatus[keyof typeof ExportMessagesStatus];


export const ExportMessagesStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export type GetLogsParams = {
level?: GetLogsLevel;
category?: string;
search?: string;
page?: number;
pageSize?: number;
};

export type GetLogsLevel = typeof GetLogsLevel[keyof typeof GetLogsLevel];


export const GetLogsLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;