# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Dispatch — iMessage Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get message history
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [sent, failed, pending]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
        - name: sort
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated message list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePage"
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message via Mac Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: Send results for each number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Mac Agent not configured or unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/export:
    get:
      operationId: exportMessages
      tags: [messages]
      summary: Export message history
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, json]
        - name: status
          in: query
          schema:
            type: string
            enum: [sent, failed, pending]
      responses:
        "200":
          description: Export file
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
...
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message from history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Settings ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [sent, failed, pending]
        error:
          type: ["string", "null"]
        agentId:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        retryCount:
          type: integer
        sentAt:
          type: string
          format: date-time
      required: [id, phoneNumber, content, status, retryCount, sentAt]
    MessagePage:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [messages, total, page, pageSize, totalPages]
    MessageInput:
      type: object
      properties:
        phoneNumbers:
          type: array
          items:
            type: string
          minItems: 1
        content:
          type: string
          minLength: 1
      required: [phoneNumbers, content]
    SendResult:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required: [sent, failed, total, results]

    # ── Settings ────────────────────────────────────────────────
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]