# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dispatch — iMessage Dashboard API
 * OpenAPI spec version: 0.2.0
 */

export interface MessageInput {
  /** @minItems 1 */
  phoneNumbers: string[];
  /** @minLength 1 */
  content: string;
}