# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Dispatch — iMessage Dashboard API
 * OpenAPI spec version: 0.2.0
 */

export interface SuccessResponse {
  success: boolean;
}